requirements.txt
matplotlib
opencv-python
numpy
tensorflow
scikit-learn
ipykernel
tk-tools
pillow